Move most of RVFI code into its own file.
This unclutters the main simulator code considerably.
